Move geo_shape queries into Tests project
This commit moves the geo_shape queries into the Test project.
Missed in the Tests refactoring
