+++ src/Tests/Tests/QueryDsl/Geo/Shape/Circle/GeoShapeCircleQueryUsageTests.cs
@@ -1,7 +1,7 @@
 ﻿using Nest;
+using Tests.Core.ManagedElasticsearch.Clusters;
 using Tests.Framework.Integration;
-using Tests.Framework.ManagedElasticsearch.Clusters;
-using Tests.Framework.MockData;
+using Tests.Domain;
 using static Nest.Infer;
 
 namespace Tests.QueryDsl.Geo.Shape.Circle
