+++ src/Tests/Tests/QueryDsl/Geo/Shape/Envelope/GeoShapeEnvelopeQueryUsageTests.cs
@@ -1,39 +1,39 @@
 ﻿using System.Collections.Generic;
 using Nest;
+using Tests.Core.ManagedElasticsearch.Clusters;
+using Tests.Domain;
 using Tests.Framework.Integration;
-using Tests.Framework.ManagedElasticsearch.Clusters;
-using Tests.Framework.MockData;
 using static Nest.Infer;
 
-namespace Tests.QueryDsl.Geo.Shape.MultiPoint
+namespace Tests.QueryDsl.Geo.Shape.Envelope
 {
-	public class GeoShapeMultiPointQueryUsageTests : GeoShapeQueryUsageTestsBase
+	public class GeoShapeEnvelopeQueryUsageTests : GeoShapeQueryUsageTestsBase
 	{
-		public GeoShapeMultiPointQueryUsageTests(ReadOnlyCluster i, EndpointUsage usage) : base(i, usage) { }
+		public GeoShapeEnvelopeQueryUsageTests(ReadOnlyCluster i, EndpointUsage usage) : base(i, usage) { }
 
 		private readonly IEnumerable<GeoCoordinate> _coordinates = new GeoCoordinate[]
 		{
-			new [] {-77.03653, 38.897676},
-			new [] {-77.009051, 38.889939 }
+			new [] { -45.0, 45.0 },
+			new [] { 45.0, -45.0 }
 		};
 
 		protected override object ShapeJson => new
 		{
-			type ="multipoint",
+			type ="envelope",
 			coordinates = this._coordinates
 		};
 
-		protected override QueryContainer QueryInitializer => new GeoShapeMultiPointQuery
+		protected override QueryContainer QueryInitializer => new GeoShapeEnvelopeQuery
 		{
 			Name = "named_query",
 			Boost = 1.1,
 			Field = Field<Project>(p=>p.Location),
-			Shape = new MultiPointGeoShape(this._coordinates),
+			Shape = new EnvelopeGeoShape(this._coordinates),
 			Relation = GeoShapeRelation.Intersects,
 		};
 
 		protected override QueryContainer QueryFluent(QueryContainerDescriptor<Project> q) => q
-			.GeoShapeMultiPoint(c => c
+			.GeoShapeEnvelope(c => c
 				.Name("named_query")
 				.Boost(1.1)
 				.Field(p=>p.Location)
@@ -41,7 +41,7 @@ namespace Tests.QueryDsl.Geo.Shape.MultiPoint
 				.Relation(GeoShapeRelation.Intersects)
 			);
 
-		protected override ConditionlessWhen ConditionlessWhen => new ConditionlessWhen<IGeoShapeMultiPointQuery>(a => a.GeoShape as IGeoShapeMultiPointQuery)
+		protected override ConditionlessWhen ConditionlessWhen => new ConditionlessWhen<IGeoShapeEnvelopeQuery>(a => a.GeoShape as IGeoShapeEnvelopeQuery)
 		{
 			q =>  q.Field = null,
 			q =>  q.Shape = null,
