+++ src/Tests/Tests/QueryDsl/Geo/Shape/GeometryCollection/GeoShapeGeometryCollectionQueryUsageTests.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using Nest;
+using Tests.Core.ManagedElasticsearch.Clusters;
+using Tests.Domain;
 using Tests.Framework.Integration;
-using Tests.Framework.ManagedElasticsearch.Clusters;
-using Tests.Framework.MockData;
 
 namespace Tests.QueryDsl.Geo.Shape.GeometryCollection
 {
