+++ src/Tests/Tests/QueryDsl/Geo/Shape/MultiLineString/GeoShapeMultiLineStringQueryUsageTests.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using Nest;
+using Tests.Core.ManagedElasticsearch.Clusters;
+using Tests.Domain;
 using Tests.Framework.Integration;
-using Tests.Framework.ManagedElasticsearch.Clusters;
-using Tests.Framework.MockData;
 using static Nest.Infer;
 
 namespace Tests.QueryDsl.Geo.Shape.MultiLineString
